Average pairs
Given a folder with rasters whose names place them in calendar order, this code will average together consecutive pairs in the list; i.e. my folder contained 183 days of data, 366 rasters, a high and low LST raster for each day of the growing season. This script was written to average them all together to produce a mean LST raster for each of the 183 days.